Got a basic call through the calendar API working

+++ Kanban.Contracts/Request/Patch/DatePatchRequest.cs
@@ -2,13 +2,13 @@
 
 public class DatePatchRequest
 {
-    public byte DateOrder { get; set; } // Is the date itself - e.g. The 15th
+    public int DateOrder { get; set; } // Is the date itself - e.g. The 15th
 
-    public byte WeekOrder { get; set; } // Is the numbered week - e.g The 2nd week
+    public int WeekOrder { get; set; } // Is the numbered week - e.g The 2nd week
 
-    public byte DayOfTheWeekOrder { get; set; } // Is the day of the week starting with Sunday as 0 - e.g. The 6th day of the week (Saturday; zero-based)
+    public int DayOfTheWeekOrder { get; set; } // Is the day of the week starting with Sunday as 0 - e.g. The 6th day of the week (Saturday; zero-based)
 
-    public byte MonthOrder { get; set; }
+    public int MonthOrder { get; set; }
 
     public string MonthName { get; set; }
